YUK-4457 -Fixes transition between residential to commercial account types and vice versa -Adds extra safety checks to make sure of residential or commercial customer type before loading customer object -Fixes problem where current user would not see the changes made to the customer if it was the active account they were viewing -Fixes problem where a commercial customer would not delete if it had no company address listed (addressID of 0) -Uncommented the dbchangemsgs for Customer Account changes.  This was preventing webservers from staying in sync in multi-server deployments when accounts were imported or edited.

+++ yukon-client/common/src/com/cannontech/database/data/customer/CICustomerBase.java
@@ -169,11 +169,11 @@ public class CICustomerBase extends Customer implements com.cannontech.common.ed
 		
 		//delete("CustomerAddress", "AddressID", getCiCustomerBase().getAddressID() );
 		getAddress().setAddressID( getCiCustomerBase().getMainAddressID() );
-//		com.cannontech.database.db.customer.CICustomerBase.getCICustomerAddressID( getCustomerID(), getDbConnection()) );
+		//com.cannontech.database.db.customer.CICustomerBase.getCICustomerAddressID( getCustomerID(), getDbConnection()) );
 	
 		getCiCustomerBase().delete();
 	
-		getAddress().delete();
+		if(getAddress().getAddressID() != 0) getAddress().delete();
 		
 		super.delete();
